requirements.txt
streamlit
google-generativeai
regex
